+++ sickrage/core/caches/name_cache.py
@@ -20,6 +20,7 @@
 from __future__ import unicode_literals
 
 import time
+import unicodedata
 from datetime import datetime, timedelta
 
 from CodernityDB.database import RecordNotFound
@@ -125,6 +126,27 @@ class NameCache(object):
             self.last_update[show.name] = datetime.fromtimestamp(int(time.mktime(datetime.today().timetuple())))
 
             self.clear(show.indexerid)
+
+            show_names = []
             for curSeason in [-1] + get_scene_seasons(show.indexerid):
                 for name in list(set(get_scene_exceptions(show.indexerid, season=curSeason) + [show.name])):
-                    self.put(name, show.indexerid)
+                    show_names.append(name)
+
+                    # strip accents
+                    try:
+                        try:
+                            name.decode('ascii')
+                        except UnicodeEncodeError:
+                            pass
+
+                        show_names.append(
+                            unicodedata.normalize('NFKD', name).encode('ASCII', 'ignore')
+                        )
+                        show_names.append(
+                            unicodedata.normalize('NFKD', name).encode('ASCII', 'ignore').replace("'", " ")
+                        )
+                    except UnicodeDecodeError:
+                        pass
+
+            for show_name in set(show_names):
+                self.put(show_name, show.indexerid)